requirements.txt
pip
requests
Flask
cassandra-driver
